Refactor page view methods to multiple files
