+++ app.py
@@ -1,62 +1,25 @@
-from flask import Flask, render_template, request, redirect
-import pymongo
-from db import get_listings_collection
-from bson.objectid import ObjectId
-from bson.errors import InvalidId
+from flask import Flask
 
+# Import page view methods from other files.
+from index import index
+from listing_create import listing_create
+from listing import listing
+from listing_update import listing_update
+from listing_ask import listing_ask
+from listing_answer import listing_answer
+
+# Create new Flask app.
 app = Flask(__name__)
-listings_collection = get_listings_collection()
 
-# TODO: this file is getting a little big. we should refactor some routes into different files
+# Instead of using annotations, link routes to methods in other files.
+app.add_url_rule('/', view_func=index)
+app.add_url_rule('/newlisting', view_func=listing_create)
+app.add_url_rule('/listing/<id>', methods=["GET", "POST"], view_func=listing)
+app.add_url_rule('/listing/<id>/update', view_func=listing_update)
+app.add_url_rule('/listing/<id>/ask', view_func=listing_ask)
+app.add_url_rule('/listing/<id>/answer', view_func=listing_answer)
 
-@app.route('/')
-def index():
-    listings_cursor = listings_collection.find()
-    # create simplified listing object for each listing in the db
-    # @bruce - I don't love the way I implemented this (because in theory the db can be very very long)
-    #          but the alternative is to rate limit/paginate. we could make a todo item for that, but this works for now
-    listings = [{ # TODO: implement a projection instead
-        "id": str(listing["_id"]), 
-        "title": listing["title"], 
-        "price": listing["price"],
-        "description": listing["description"],
-        "author": listing["author"]
-        } for listing in list(listings_cursor.sort("timestamp", pymongo.DESCENDING))]
-    listings_cursor.close()
-    return render_template('index.html', listings = listings)
-
-@app.route('/newlisting')
-def listing_create():
-    return render_template('newlisting.html')
-
-@app.route('/listing/<id>', methods = ["GET", "POST"])
-def listing(id):
-    if request.method == "GET":
-        # display listing page
-        try:
-            listing_cursor = listings_collection.find({'_id': ObjectId(id)})
-            listing = listing_cursor.next() # get the first (only) listing with the id
-            listing_cursor.close()
-            questions = listing["questions"]
-            # TODO: add image support
-            return render_template('listing.html', listing = listing, questions = questions)
-        except StopIteration:
-            return "No listing found with this id."
-        except InvalidId:
-            return "The given id is invalid."
-
-@app.route('/listing/<id>/update')
-def listing_update(id):
-    # TODO: put link to answering a question in here
-    return render_template('listing_update.html')
-
-@app.route('/listing/<id>/ask')
-def listing_ask(id):
-    return render_template('listing_ask.html')
-
-@app.route('/listing/<id>/answer/<qid>')
-def listing_answer(id, qid):
-    return render_template('listing_answer.html')
-
+# Run the app.
 if __name__ == '__main__':
-    app.run(port=5000)
+    # Host 0.0.0.0 will allow you to access the web server from other computers (i.e. a phone) connected to your local network.
+    app.run(host='0.0.0.0', port=5000)
