+++ listing_answer.py
@@ -1,0 +1,4 @@
+from flask import render_template
+
+def listing_answer(id, qid):
+    return render_template('listing_answer.html')
