+++ listing_ask.py
@@ -1,0 +1,4 @@
+from flask import render_template
+
+def listing_ask(id):
+    return render_template('listing_ask.html')
