+++ listing_create.py
@@ -1,0 +1,4 @@
+from flask import render_template
+
+def listing_create():
+    return render_template('newlisting.html')
