+++ listing_update.py
@@ -1,0 +1,5 @@
+from flask import render_template
+
+def listing_update(id):
+    # TODO: put link to answering a question in here
+    return render_template('listing_update.html')
